Move Netlify & Zeit components into it's own folder
