+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/index.tsx
@@ -10,15 +10,15 @@ import getWorkspaceItems, {
   getDisabledItems,
 } from 'app/store/modules/workspace/items';
 
+import ConfigurationFiles from './items/ConfigurationFiles';
+import { Deployment } from './items/Deployment';
 import Files from './items/Files';
 import { GitHub } from './items/GitHub';
-import Server from './items/Server';
 import Live from './items/Live';
 import { More } from './items/More';
-import { Deployment } from './items/Deployment';
-import ConfigurationFiles from './items/ConfigurationFiles';
 import { NotOwnedSandboxInfo } from './items/NotOwnedSandboxInfo';
 import { ProjectInfo } from './items/ProjectInfo';
+import Server from './items/Server';
 
 import { Advertisement } from './Advertisement';
 import Chat from './Chat';
