+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/items/Deployment/Netlify/index.ts
@@ -1,0 +1,1 @@
+export { Netlify } from './Netlify';
