+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/items/Deployment/Zeit/index.ts
@@ -1,0 +1,1 @@
+export { Zeit } from './Zeit';
